querys fixed

+++ app/db/crud/agency_crud.py
@@ -1,7 +1,6 @@
 from fastapi import HTTPException, status
-from sqlalchemy import delete
 from sqlalchemy.orm import Session
-from models import AgencyModel, PackageReservation, AgencyExcursionAssociation, ExcursionReservation, AgencyOfferAssociation, PackageModel
+from models import AgencyModel, PackageReservation, AgencyExcursionAssociation, ExcursionReservation, PackageModel
 from schemas import AgencySchema
 
 
@@ -21,28 +20,25 @@ def create_agency(db: Session, agency_create: AgencySchema):
     return "Success"
 
 def delete_agency(db: Session, agency_delete_id: int):
-    print(agency_delete_id)
-    agency = get_agency(db, agency_delete_id)
 
+    agency = get_agency(db, agency_delete_id)
     if agency is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Agency not found")
     
-    packages_with_agency = db.query(PackageModel).filter(PackageModel.agency_id == agency_delete_id).all()
-    if packages_with_agency is not None:
-        for package in packages_with_agency:
-            reserved_package = db.query(PackageReservation).filter(PackageReservation.package_id == package.id).first()
-            if reserved_package is not None:
-                raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't delete this Agency because is is ivolved in a Package Reservation")
-    
-    excursions_associated_with_agency = db.query(AgencyExcursionAssociation).filter(AgencyExcursionAssociation.agency_id == agency_delete_id).all()
+    reserved_package_with_agency = db.query(PackageModel).\
+        join(PackageReservation, PackageModel.id == PackageReservation.package_id).\
+        filter(PackageModel.agency_id == agency_delete_id).first()
 
-    for excursion in excursions_associated_with_agency:
-        print(excursion)
-        excursion_reservation = db.query(ExcursionReservation).filter(ExcursionReservation.excursion_id == excursion.excursion_id).first()
-        if excursion_reservation is not None:
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't delete this Agency because is is ivolved in an Excursion Reservation")
-    
+    if reserved_package_with_agency is not None:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't delete this Agency because is is ivolved in a Package Reservation")
 
+    reserved_excursion_with_agency = db.query(AgencyExcursionAssociation).\
+        join(ExcursionReservation, AgencyExcursionAssociation.excursion_id == ExcursionReservation.excursion_id).\
+        filter(AgencyExcursionAssociation.agency_id == agency_delete_id).first()
+        
+    if reserved_excursion_with_agency is not None:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't delete this Agency because is is ivolved in an Excursion Reservation")
+    
     db.delete(agency)
     db.commit()
 
@@ -54,21 +50,20 @@ def update_agency(db: Session, agency_update: AgencySchema):
     if agency is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Agency not found")
     
-    packages_with_agency = db.query(PackageModel).filter(PackageModel.agency_id == agency_update.id).all()
-    if packages_with_agency is not None:
-        for package in packages_with_agency:
-            reserved_package = db.query(PackageReservation).filter(PackageReservation.package_id == package.id).first()
-            if reserved_package is not None:
-                raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't update this Agency because is is ivolved in a Package Reservation")
-    
-    excursions_associated_with_agency = db.query(AgencyExcursionAssociation).filter(AgencyExcursionAssociation.agency_id == agency_update.id).all()
+    reserved_package_with_agency = db.query(PackageModel).\
+        join(PackageReservation, PackageModel.id == PackageReservation.package_id).\
+        filter(PackageModel.agency_id == agency_update.id).first()
 
-    for excursion in excursions_associated_with_agency:
-        print(excursion)
-        excursion_reservation = db.query(ExcursionReservation).filter(ExcursionReservation.excursion_id == excursion.excursion_id).first()
-        if excursion_reservation is not None:
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't update this Agency because is is ivolved in an Excursion Reservation")
-    
+    if reserved_package_with_agency is not None:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't update this Agency because is is ivolved in a Package Reservation")
+
+    reserved_excursion_with_agency = db.query(AgencyExcursionAssociation).\
+        join(ExcursionReservation, AgencyExcursionAssociation.excursion_id == ExcursionReservation.excursion_id).\
+        filter(AgencyExcursionAssociation.agency_id == agency_update.id).first()
+        
+    if reserved_excursion_with_agency is not None:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't update this Agency because is is ivolved in an Excursion Reservation")
+  
     if agency_update.name is not None:
         agency.name = agency_update.name
     if agency_update.address is not None:
@@ -86,7 +81,6 @@ def update_agency(db: Session, agency_update: AgencySchema):
 
 def toModel(schema:AgencySchema) -> AgencyModel:
     return AgencyModel(
-        # id=schema.id, 
                        name=schema.name, 
                        address=schema.address, 
                        fax_number=schema.fax_number,
